feat: create login modal
