+++ apps/web/src/components/Modals/LoginModal/index.ts
@@ -1,0 +1,3 @@
+import LoginModal from './LoginModal';
+
+export default LoginModal;
